// src/api/interfaces/contingency.interfaces.d.ts
// Generated by https://quicktype.io
export interface ContingenciesTmHttp {
  docs: ContingencyHttp[];
  totalDocs: number;
  limit: number;
  totalPages: number;
  page: number;
  pagingCounter: number;
  hasPrevPage: boolean;
  hasNextPage: boolean;
  prevPage: null;
  nextPage: number;
}

export interface ContingencyHttp {
  _id: string;
  folio: string;
  id_employee: number;
  name_employee: string;
  date: string;
  half_day: boolean;
  status: string;
  comments: string;
  observations: string;
  id_tm: number;
  createdAt: string;
  updatedAt: string;
  __v: number;
}